+++ src/mpid/ch3/src/mpidi_isend_self.c
@@ -74,7 +74,7 @@ int MPIDI_Isend_self(const void * buf, int count, MPI_Datatype datatype, int ran
 	    
 	MPIDI_CH3U_Buffer_copy(buf, count, datatype, &sreq->status.MPI_ERROR,
 			       rreq->dev.user_buf, rreq->dev.user_count, rreq->dev.datatype, &data_sz, &rreq->status.MPI_ERROR);
-	MPIR_STATUS_SET_COUNT(rreq->status, (int)data_sz);
+	MPIR_STATUS_SET_COUNT(rreq->status, data_sz);
 	MPID_REQUEST_SET_COMPLETED(rreq);
 	MPID_Request_release(rreq);
 	/* sreq has never been seen by the user or outside this thread, so it is safe to reset ref_count and cc */
